+++ es6/admin-event-espresso.js
@@ -1,8 +1,15 @@
 /**
-* ask browsers to stop with the autocorrupt already
-*/
-jQuery("input.eway-no-autocorrupt").attr({
-	autocorrect:	"off",
-	autocapitalize:	"off",
-	spellcheck:		"false"
-});
+ * ask browsers to stop with the autocorrupt already
+ */
+(function() {
+
+	const fields = document.querySelectorAll(".eway-no-autocorrupt");
+
+	for (let i = 0, len = fields.length; i < len; i++) {
+		const field = fields[i];
+		field.autocorrect		= field.type === "password" ? "new-password" : "off";
+		field.autocapitalize	= "off";
+		field.spellcheck		= false;
+	}
+
+})();