+++ includes/integrations/event_espresso_eway/EE_PMT_event_espresso_eway.pm.php
@@ -96,6 +96,7 @@ class EE_PMT_event_espresso_eway extends \EE_PMT_Base {
 	* @return EE_Payment_Method_Form
 	*/
 	public function generate_new_settings_form() {
+		// NB: class names have a trailing space because EE appends additional classes without a space!
 
 		$form = new \EE_Payment_Method_Form([
 
@@ -103,32 +104,32 @@ class EE_PMT_event_espresso_eway extends \EE_PMT_Base {
 
 				'eway_api_key'				=>	new EE_Text_Input([
 													'html_label_text'	=> _x('API key', 'settings field', 'eway-payment-gateway'),
-													'html_class'		=> 'eway-no-autocorrupt',
+													'html_class'		=> 'eway-no-autocorrupt ',
 												]),
 
-				'eway_password'				=>	new EE_Text_Input([
+				'eway_password'				=>	new EE_Password_Input([
 													'html_label_text'	=> _x('API password', 'settings field', 'eway-payment-gateway'),
-													'html_class'		=> 'eway-no-autocorrupt',
+													'html_class'		=> 'eway-no-autocorrupt ',
 												]),
 
 				'eway_ecrypt_key'			=>	new EE_Text_Area_Input([
 													'html_label_text'	=> _x('Client Side Encryption key', 'settings field', 'eway-payment-gateway'),
-													'html_class'		=> 'eway-no-autocorrupt',
+													'html_class'		=> 'eway-no-autocorrupt ',
 												]),
 
 				'eway_sandbox_api_key'		=>	new EE_Text_Input([
 													'html_label_text'	=> _x('Sandbox API key', 'settings field', 'eway-payment-gateway'),
-													'html_class'		=> 'eway-no-autocorrupt',
+													'html_class'		=> 'eway-no-autocorrupt ',
 												]),
 
-				'eway_sandbox_password'		=>	new EE_Text_Input([
+				'eway_sandbox_password'		=>	new EE_Password_Input([
 													'html_label_text'	=> _x('Sandbox API password', 'settings field', 'eway-payment-gateway'),
-													'html_class'		=> 'eway-no-autocorrupt',
+													'html_class'		=> 'eway-no-autocorrupt ',
 												]),
 
 				'eway_sandbox_ecrypt_key'	=>	new EE_Text_Area_Input([
 													'html_label_text'	=> _x('Sandbox Client Side Encryption key', 'settings field', 'eway-payment-gateway'),
-													'html_class'		=> 'eway-no-autocorrupt',
+													'html_class'		=> 'eway-no-autocorrupt ',
 												]),
 
 				'eway_logging'				=>	new EE_Select_Input([
@@ -160,7 +161,7 @@ class EE_PMT_event_espresso_eway extends \EE_PMT_Base {
 		if ($plugin_page === 'espresso_payment_settings') {
 			$min = SCRIPT_DEBUG ? '' : '.min';
 			$ver = SCRIPT_DEBUG ? time() : EWAY_PAYMENTS_VERSION;
-			wp_enqueue_script('event_espresso_eway-settings', plugins_url("js/admin-event-espresso$min.js", EWAY_PAYMENTS_PLUGIN_FILE), ['jquery'], $ver, true);
+			wp_enqueue_script('event_espresso_eway-settings', plugins_url("js/admin-event-espresso$min.js", EWAY_PAYMENTS_PLUGIN_FILE), [], $ver, true);
 		}
 	}